pyproject.toml
[project]
name = "naneos-upload-tray"
version = "0.1.0"
description = "Naneos Manager running in the system tray."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "naneos-devices>=1.1.6",
    "pillow>=11.3.0",
    "pystray>=0.19.5",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pyinstaller>=6.16.0",
    "ruff>=0.13.1",
]
